refactor: Move SVG components to artworks directory
