+++ components/svg/artworks/book-reading.tsx
