+++ components/svg/artworks/projects.tsx
